# Speed of gevent.local.local more, especially for subclasses (2-3x faster)