@jamadden 2018-02-23
For fun, here's a look at the difference compiling with Cython 0.28 makes. First, the code prior to this commit (the code in 1.3a1):

| Benchmark          | lCython 0.27.3 | Cython 0.28            |
|--------------------|------------------|-----------------------------|
| getattr gevent     | 215 ns        | 190 ns: 1.13x faster (-11%) |
| setattr gevent     | 212 ns        | 180 ns: 1.18x faster (-15%) |
| getattr gevent sub | 563 ns        | 540 ns: 1.04x faster (-4%)  |
| setattr gevent sub | 558 ns        | 528 ns: 1.06x faster (-5%)  |

For even more fun, let's look at the gevent 1.2.2 code (uncompiled) versus this branch with Cython 0.28:

| Benchmark          | local_36_122 | local_36_tweak2              |
|--------------------|------------------|-----------------------------|
| getattr gevent     | 6.62 us      | 158 ns: 41.88x faster (-98%) |
| setattr gevent     | 7.27 us      | 165 ns: 44.15x faster (-98%) |
| getattr gevent sub | 7.15 us      | 175 ns: 40.90x faster (-98%) |
| setattr gevent sub | 7.06 us      | 179 ns: 39.45x faster (-97%) |